+++ Presentation(UI)/ViewModels/SummaryViewModel.cs
@@ -4,7 +4,7 @@
     {
         public int SummaryID { get; set; }
         public int ProjectID { get; set; }
-        public int TotalHours { get; set; }
+        public decimal TotalHours { get; set; }
         public decimal TotalPrice { get; set; }
         public string Notes { get; set; }
     }
